Moved implementation of MessageData to proc macros
This should make it very easy to implement the protocol and make it more
open for changes in the future.

There are still some quirks that make it so that this patch does not
work properly. The macros still need some work.

Most notably the chat might not work yet.