Impl. subject, object and chat getters in macro
This commit adds additional attributes that adapt the getters of the
MessageDataV2 trait. This way the protocol has a way to define which
field is the subject player, object player, and the target chat.

This goes in line with the efforts to rewrite the MessageData enum to
contain single structs, that implement a common trait.